+++ src/hooks/usePageTitle.ts
@@ -12,10 +12,8 @@ const routeTitles: Record<string, string> = {
   '/schedule': 'Sessions',
   '/take-attendance': 'Take Attendance',
   '/accounts': 'Accounts',
-  '/officer-accounts': 'Officer Accounts',
   '/excuse-application': 'Excuse Application',
   '/academic-year': 'Allowed Terms',
-  '/subjects': 'Subjects',
   '/login': 'Login',
   '/sessions': 'Session Students',
   '/model-training-signature-classify': 'Model Training',
